+++ java/integTest/src/test/java/glide/TransactionTestUtilities.java
@@ -95,6 +95,8 @@ public class TransactionTestUtilities {
         // TODO
         // baseTransaction.pfadd(hllKey1, new String[0]);
 
+        // TODO pfcount
+
         return baseTransaction;
     }