+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Core: Enabled Cluster Mode periodic checks by default ([#1089](https://github.com/aws/glide-for-redis/pull/1089))
 * Node: Added Rename command. ([#1124](https://github.com/aws/glide-for-redis/pull/1124))
 * Core: Added `PFADD` command. (TODO ADD HERE PR #)
+* Core: Added `PFCOUNT` command. (TODO ADD HERE PR #)
 
 #### Features